style: simply solve the circular dependency problem of allocator and canvas
